Revert the deletion of AbstractTask
