+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/TaskDestroyablesInternal.java
@@ -22,7 +22,7 @@ import org.gradle.api.tasks.TaskDestroyables;
 
 
 /**
- * Note: this is currently not visible on {@link org.gradle.api.internal.TaskInternal} to avoid it leaking onto {@link org.gradle.api.DefaultTask} and so on to the public API.
+ * Note: this is currently not visible on {@link org.gradle.api.internal.TaskInternal} to avoid it leaking onto {@link org.gradle.api.internal.AbstractTask} and so on to the public API.
  */
 public interface TaskDestroyablesInternal extends TaskDestroyables {
