+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/ExecutionGlobalServices.java
@@ -123,6 +123,8 @@ public class ExecutionGlobalServices {
     );
 
     TypeAnnotationMetadataStore createAnnotationMetadataStore(CrossBuildInMemoryCacheFactory cacheFactory) {
+        @SuppressWarnings("deprecation")
+        Class<?> deprecatedAbstractTask = org.gradle.api.internal.AbstractTask.class;
         return new DefaultTypeAnnotationMetadataStore(
             ImmutableSet.of(
                 CacheableTask.class,
@@ -135,6 +137,7 @@ public class ExecutionGlobalServices {
                 "kotlin"
             ),
             ImmutableSet.of(
+                deprecatedAbstractTask,
                 ClosureBackedAction.class,
                 ConfigureUtil.WrappedConfigureAction.class,
                 ConventionTask.class,
