+++ subprojects/core/src/test/groovy/org/gradle/api/internal/DefaultTaskTest.groovy
@@ -60,7 +60,7 @@ class DefaultTaskTest extends AbstractTaskTest {
     def "default task"() {
         given:
         def identity = TaskIdentity.create(TEST_TASK_NAME, Task, project)
-        Task task = DefaultTask.injectIntoNewInstance(project, identity, { new DefaultTask() } as Callable)
+        Task task = AbstractTask.injectIntoNewInstance(project, identity, { new DefaultTask() } as Callable)
 
         expect:
         task.dependsOn.isEmpty()
@@ -76,7 +76,7 @@ class DefaultTaskTest extends AbstractTaskTest {
     def "can inject values into task when using no-args constructor"() {
         given:
         def identity = TaskIdentity.create(TEST_TASK_NAME, Task, project)
-        def task = DefaultTask.injectIntoNewInstance(project, identity, { new DefaultTask() } as Callable)
+        def task = AbstractTask.injectIntoNewInstance(project, identity, { new DefaultTask() } as Callable)
 
         expect:
         task.project.is(project)
