+++ subprojects/core/src/testFixtures/groovy/org/gradle/api/tasks/AbstractSpockTaskTest.groovy
@@ -21,6 +21,7 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Project
 import org.gradle.api.Task
+import org.gradle.api.internal.AbstractTask
 import org.gradle.api.internal.project.ProjectInternal
 import org.gradle.api.internal.project.taskfactory.AnnotationProcessingTaskFactory
 import org.gradle.api.internal.project.taskfactory.DefaultTaskClassInfoStore
@@ -128,7 +129,7 @@ abstract class AbstractSpockTaskTest extends AbstractProjectBuilderSpec {
         getTask().actions = [action2]
 
         then:
-        [new DefaultTask.TaskActionWrapper(action2, "doLast(Action)")] ==  getTask().actions
+        [new AbstractTask.TaskActionWrapper(action2, "doLast(Action)")] ==  getTask().actions
     }
 
     def testAddActionWithNull() {
