+++ subprojects/language-native/src/integTest/groovy/org/gradle/language/AbstractNativeParallelIntegrationTest.groovy
@@ -93,6 +93,8 @@ abstract class AbstractNativeParallelIntegrationTest extends AbstractInstalledTo
         setupParallelTaskAndExpectations(taskName)
 
         buildFile << """
+            import org.gradle.api.internal.AbstractTask
+
             ${callbackWorkerLeaseService}
 
             tasks.matching { it.name == '${taskName}' }.all { task ->
